significant touchups and documentation
