+++ Localization/en-US/Mods.PathOfTerraria.Misc.hjson
@@ -7,7 +7,7 @@ GrimoireConsume: "{0} consumed!"
 GrimoireMaterial: Used for the Grimoire's Menagerie.
 GrimoireBoon: Traded {0} for coin!
 # This is used for disabling EoW/BoC spawns, instead of spawning the boss it fails and says this.
-EvilBossFailedToSummon: The distant screams fade.
+ChasmAppears: A chasm appears...
 HealthPotionTooltip: Health Potion
 ManaPotionTooltip: Mana Potion
 RestoresHealthTooltip: Restores {0} health
